+++ src/main/java/com/openmatching/engine/Side.java
@@ -1,4 +1,4 @@
-package org.naveedh;
+package com.openmatching.engine;
 
 public enum Side {
     BUY, SELL;
