--- src/main/java/org/naveedh/Main.java
@@ -1,7 +1,0 @@
-package org.naveedh;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
