--- src/main/java/org/naveedh/PQObject.java
@@ -1,11 +1,0 @@
-package org.naveedh;
-
-public class PQObject {
-    int price;
-    Order order;
-
-    public PQObject(int price, Order order) {
-        this.price = price;
-        this.order = order;
-    }
-}
