+++ src/main/resources/scenario1.txt
@@ -1,0 +1,8 @@
+BUY, 150, 20
+BUY, 100, 50
+SELL, 200, 25
+SELL, 250, 100
+BUY, 210, 25
+SELL, 50, 100
+BUY, 150, 20
+BUY, 100, 50
