+++ src/main/resources/scenario2.txt
@@ -1,0 +1,8 @@
+BUY, 100, 100
+BUY, 105, 100
+BUY, 110, 100
+BUY, 115, 100
+BUY, 120, 100
+SELL, 100, 385
+BUY, 100, 500
+SELL, 100, 1850
