+++ src/main/resources/scenario3.txt
@@ -1,0 +1,29 @@
+BUY, 109, 1
+BUY, 110, 1
+BUY, 111, 1
+BUY, 112, 1
+BUY, 113, 1
+BUY, 114, 1
+BUY, 115, 1
+BUY, 116, 1
+BUY, 117, 1
+BUY, 118, 1
+BUY, 119, 1
+BUY, 120, 1
+BUY, 121, 1
+BUY, 122, 1
+SELL, 109, 1
+SELL, 111, 1
+SELL, 113, 1
+SELL, 114, 1
+SELL, 115, 1
+SELL, 116, 1
+SELL, 117, 1
+SELL, 118, 1
+SELL, 119, 1
+SELL, 120, 1
+SELL, 121, 1
+SELL, 122, 1
+SELL, 123, 1
+SELL, 124, 1
+SELL, 125, 1
