+++ src/main/resources/scenario4.txt
@@ -1,0 +1,10 @@
+BUY, 100, 50
+SELL, 105, 30
+BUY, 102, 40
+SELL, 101, 20
+BUY, 103, 60
+SELL, 100, 70
+SELL, 99, 50
+BUY, 98, 80
+SELL, 97, 30
+BUY, 105, 40
