+++ src/main/resources/scenario5.txt
@@ -1,0 +1,6 @@
+BUY, 100, 50
+SELL, 105, 30
+BUY, 110, 40
+SELL, 100, 20
+BUY, 105, 30
+SELL, 105, 25
