+++ src/main/resources/scenario6.txt
@@ -1,0 +1,5 @@
+BUY, 100, 50
+SELL, 90, 30
+SELL, 95, 25
+BUY, 92, 40
+SELL, 90, 50
